Refactor tier module rewards logic into BeginBlocker, update tests
